feat: implement mass
